Disable test parallelization
This stops the confusing intermittent functional test failures which (I'm guessing) are caused by parallel test runs in VS killing off the test Redis process prematurely.

I'd assume there is a much better way to fix this, i.e. some kind of locking mechanism or pooling for the test Redis process so that the tests can run in parallel without affecting each other?

Anyway, this works for now: it takes a bit longer to run the full suite, but all of the tests do consistently pass.